refactor(core): Move default typesetter to proper namespace
